// src/features/Scenario/types.ts
export enum SubmissionStatus {
  SUBMISSION_STATUS_UNKNOWN = 0,
  SUBMISSION_STATUS_IN_PROGRESS = 1,
  SUBMISSION_STATUS_FAILED = 2,
  SUBMISSION_STATUS_SUCCESS = 3,
}

export enum PublicationStatus {
  PUBLICATION_STATUS_UNKNOWN = 0,
  PUBLICATION_STATUS_DRAFT = 1,
  PUBLICATION_STATUS_PUBLISHED = 2,
}

export interface BaseData {
  created_at: Timestamp;
  updated_at: Timestamp;
}

export interface Answer {
  question_id: number;
  answer: string;
  relevance: number;
  clarity_completeness: number;
  accuracy: number;
  overall: number;
  status: SubmissionStatus;
  base_data: BaseData;
}

export interface Attempt {
  id: number;
  scenario_id: number;
  answers: Answer[];
  base_data: BaseData;
  attempt_number: number;
}

export interface Submission {
  id: number;
  scenario_id: number;
  candidate_id: number;
  attempts: Attempt[];
}

export interface Field {
  id: number;
  name: string;
  base_data: BaseData;
}

export interface Scenario {
  id: number;
  name: string;
  description: string;
  base_data: BaseData;
  rating: number;
  fields: Field[];
  total_participant: number;
  questions: Question[];
}

export interface Question {
  id: number;
  criteria: string;
  hint: string;
  content: string;
  base_data: BaseData;
}